refactor(iocuak): move isContainerModuleTaskKind to domain folder
